+++ evlist.php
@@ -34,7 +34,7 @@
 *   @copyright  Copyright (c) 2008 - 2010 Mark R. Evans mark AT glfusion DOT org
 *   @copyright  Copyright (c) 2010 - 2017 Lee Garner <lee@leegarner.com>
 *   @package    evlist
-*   @version    1.4.1
+*   @version    1.4.2
 *   @license    http://opensource.org/licenses/gpl-2.0.php
 *               GNU Public License v2 or later
 *   @filesource
@@ -51,7 +51,7 @@ if (!defined ('GVERSION')) {
 */
 $_EV_CONF['pi_name']         = 'evlist';
 $_EV_CONF['pi_display_name'] = 'Event Calendar';
-$_EV_CONF['pi_version']      = '1.4.1';
+$_EV_CONF['pi_version']      = '1.4.2';
 $_EV_CONF['gl_version']      = '1.4.0';
 $_EV_CONF['pi_url']          = 'http://www.glfusion.org';