+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.4.1
 - Add comment support
+- Moved ticket list to new tab on event page
 
 ## v1.4.0
 - Add event signups